requirements.txt
Flask==2.0.3
flask-cors==5.0.1
python-dotenv==0.19.2
requests==2.26.0
numpy==1.21.2
gunicorn==23.0.0